skfilter: display cutoff and notify clamping
Sometimes, the filter is enforced to stay in its stable range (to prevent it from blowing up). In this case, the cutoff at which the filter remains stable is lower than what the user specified. In that case, show a warning with an indication on how to fix the problem.